Updated LPC Flash drivers.
Added bank selection to the LPC43xx Flash driver.
Fixed handling of non-uniform sectors on LPC11Exx.
Fixed position handling in EEPROM driver.
Minor update in MMCSD driver, changed types of IF_POSITION param writing.
Added Flash Geometry API.

+++ generic/mmcsd.c
@@ -1140,14 +1140,14 @@ static enum Result cardSetParam(void *object, int parameter, const void *data)
 
     case IF_POSITION:
     {
-      const uint64_t position = (uint64_t)(*(const uint32_t *)data);
+      const uint32_t position = *(const uint32_t *)data;
 
       /* Check address alignment */
       if (position & MASK(BLOCK_POW))
         return E_VALUE;
 
       /* Check address range */
-      if ((position >> BLOCK_POW) >= (uint64_t)device->info.sectorCount)
+      if ((position >> BLOCK_POW) >= device->info.sectorCount)
         return E_VALUE;
 
       device->transfer.position = position;
